refactor(fwstate): extract dataplane test's CGo harness into own files
